+++ src/test/javascript/jsdoc_test.ts
@@ -28,6 +28,14 @@ suite('jsdoc', () => {
       });
     });
 
+    test('parses CRLF comments', () => {
+      const parsed = jsdoc.parseJsdoc('* Just some text\r\n* in multiple lines.');
+      assert.deepEqual(parsed, {
+        description: 'Just some text\nin multiple lines.',
+        tags: [],
+      });
+    });
+
     test('parses body-only', () => {
       const parsed = jsdoc.parseJsdoc('* Just some text\n* in multiple lines.');
       assert.deepEqual(parsed, {